Create tests (WIP)
