+++ tests/notebooks/test_hhl_lanchester.py
@@ -1,0 +1,11 @@
+from tests.utils_for_testbook import (
+    validate_quantum_program_size,
+    validate_quantum_model,
+    wrap_testbook,
+)
+from testbook.client import TestbookNotebookClient
+
+
+@wrap_testbook("hhl_lanchester", timeout_seconds=450)
+def test_notebook(tb: TestbookNotebookClient) -> None:
+    pass  # TODO
